MAINT: Move SMIRNOFF code into new submodule
